Refactored project Since it is a simple script I just compile on the command line, I removed all visual studio stuff. Thus, I refactored the file structure a bit. In addition I - added an option to download localized image for specific country, can be set in app's config file. - moved call of DownloadJson to main and changed/tweaked other minor things.
