+++ BingBackground/Properties/AssemblyInfo.cs
