+++ BingBackground/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+	<package id="Newtonsoft.Json" version="6.0.8" targetFramework="net45" />
+</packages>
