+++ BingBackground/packages/repositories.config
@@ -1,4 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <repositories>
-  <repository path="..\BingBackground\packages.config" />
+	<repository path="..\packages.config" />
 </repositories>
