+++ README.md
@@ -2,12 +2,12 @@
 BingBackground is a small command-line program that downloads the Bing homepage image and sets it as your desktop background. Images are saved in your picture directory in a folder named "Bing Backgrounds".
 
 ## Usage
-Clone the GitHub repo and run the executable found [here](https://github.com/josueespinosa/BingBackground/tree/master/BingBackground/BingBackground/bin/Release), or download the executable and dll file *only* from [Google Drive](https://drive.google.com/folderview?id=0B8bo8ihOy3gLfjJ2R1dxU0ZDU0hOc1ZJdTFHd0dHN0g1VjlueWV0OWF5YWNFa0pQX1E3WTg&usp=sharing).
+Clone the GitHub repo and run the executable found [here](https://github.com/SunRed/BingBackground/tree/master/BingBackground/bin).
 
 ## Source
-The source code currently consists of one .cs file located [here](https://github.com/josueespinosa/BingBackground/blob/master/BingBackground/BingBackground/BingBackground.cs). Contributions are always welcome!
+The source code currently consists of one .cs file located [here](https://github.com/SunRed/BingBackground/tree/master/BingBackground/BingBackground.cs). Contributions are always welcome!
 
-## Author
+## Original Author
 Josue Espinosa, josue.espinosa.live@gmail.com
 
 ## License
